Implement Activity edit/delete feature in Activity space
